backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.11
python-dotenv==1.0.1
pandas==2.2.2
